+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
+## v0.7.6
+### New Features
+#### Living assigns per process cache
+The performance of `poke` and `peek` operations increases significantly, as the event handler
+does not have to get the assigns from the browser each time.
+
 ## v0.7.5
 ### New Features
 #### Living assigns inside the shared comamander